updateSignature

Update a signature in the database

Usage

updateSignature(conn_handler, signature_id, omic_signature, verbose = TRUE)

Arguments

conn_handler

An established connection handler using SigRepo::newConnhandler()

signature_id

a unique signature id in the database that needs to be updated

omic_signature

An R6 class object from the OmicSignature package

verbose

a logical value indicates whether or not to print the diagnostic messages. Default is TRUE.

Examples

# Create a db connection
conn_handler <- SigRepo::newConnHandler(
 dbname = "sigrepo", 
 host = "montilab.bu.edu", 
 port = 3306, 
 user = "guest", 
 password = "guest"
)

# Get a list of signatures that belongs to user = 'guest'
signature_tbl <- SigRepo::searchSignature(
 conn_handler = conn_handler,
 user_name = "guest"
)
#> Error in value[[3L]](cond): Error in .local(drv, ...): Failed to connect to database: Error: Can't connect to MySQL server on 'montilab.bu.edu:3306' (111)
#> 
#> 

# Update the desired signature (NOT RUN)
# SigRepo::updateSignature(
#  conn_handler = conn_handler,
#  signature_id = signature_tbl$signature_id,
#  omic_signature = omic_signature 
# )