addSignature

Add signature to database

Usage

addSignature(
  conn_handler,
  omic_signature,
  visibility = FALSE,
  return_signature_id = FALSE,
  return_missing_features = FALSE,
  verbose = TRUE
)

Arguments

conn_handler

A handler uses to establish connection to the database obtained from SigRepo::newConnhandler() (required)

omic_signature

An R6 class object from OmicSignature package (required)

visibility

A logical value indicates whether or not to allow others to view and access one's uploaded signature. Default is FALSE.

return_signature_id

A logical value indicates whether or not to return the ID of the uploaded signature. Default is FALSE.

return_missing_features

A logical value indicates whether or not to return a list of features that does not exist in the database. Default is FALSE.

verbose

A logical value indicates whether or not to print the diagnostic messages. Default is TRUE.